demo/rpc_service/requirements.txt
boto3==1.9.210
grpcio==1.21.1
grpcio-tools==1.21.1
jieba